Uploaded MisMatch Game created using python
I used Python to create a mismatch game by designing a simple program that displays a set of cards with hidden values. The user has to match pairs of cards by flipping them over, and the game checks if the selected cards match or not. I utilized Python's random module to shuffle the cards and implemented loops and conditionals to control the game flow and provide feedback to the player.